requirements_test_all.txt
# Home Assistant test
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version
asynctest==0.12.1
coveralls==1.2.0
flake8-docstrings==1.0.3
flake8==3.5
mock-open==1.3.1
mypy==0.590
pydocstyle==1.1.1
pylint==1.9.2
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
pytest-sugar==0.9.1
pytest-timeout==1.3.0
pytest==3.6.3
requests_mock==1.5


# homeassistant.components.homekit
HAP-python==2.2.2

# homeassistant.components.notify.html5
PyJWT==1.6.0

# homeassistant.components.sonos
SoCo==0.14

# homeassistant.components.device_tracker.automatic
aioautomatic==0.6.5

# homeassistant.components.emulated_hue
# homeassistant.components.http
aiohttp_cors==0.7.0

# homeassistant.components.hue
aiohue==1.5.0

# homeassistant.components.notify.apns
apns2==0.3.0

# homeassistant.components.calendar.caldav
caldav==0.5.0

# homeassistant.components.sensor.coinmarketcap
coinmarketcap==5.0.3

# homeassistant.components.device_tracker.upc_connect
defusedxml==0.5.0

# homeassistant.components.sensor.dsmr
dsmr_parser==0.11

# homeassistant.components.sensor.season
ephem==3.7.6.0

# homeassistant.components.climate.honeywell
evohomeclient==0.2.5

# homeassistant.components.feedreader
# homeassistant.components.sensor.geo_rss_events
feedparser==5.2.1

# homeassistant.components.sensor.foobot
foobot_async==0.3.1

# homeassistant.components.tts.google
gTTS-token==1.1.1

# homeassistant.components.ffmpeg
ha-ffmpeg==1.9

# homeassistant.components.sensor.geo_rss_events
haversine==0.4.5

# homeassistant.components.mqtt.server
hbmqtt==0.9.2

# homeassistant.components.binary_sensor.workday
holidays==0.9.5

# homeassistant.components.frontend
home-assistant-frontend==20180710.0

# homeassistant.components.homematicip_cloud
homematicip==0.9.6

# homeassistant.components.influxdb
# homeassistant.components.sensor.influxdb
influxdb==5.0.0

# homeassistant.components.dyson
libpurecoollink==0.4.2

# homeassistant.components.media_player.soundtouch
libsoundtouch==0.7.2

# homeassistant.components.sensor.mfi
# homeassistant.components.switch.mfi
mficlient==0.3.0

# homeassistant.components.binary_sensor.trend
# homeassistant.components.image_processing.opencv
numpy==1.14.5

# homeassistant.components.mqtt
# homeassistant.components.shiftr
paho-mqtt==1.3.1

# homeassistant.components.device_tracker.aruba
# homeassistant.components.device_tracker.asuswrt
# homeassistant.components.device_tracker.cisco_ios
# homeassistant.components.device_tracker.unifi_direct
# homeassistant.components.media_player.pandora
pexpect==4.6.0

# homeassistant.components.pilight
pilight==0.1.1

# homeassistant.components.sensor.mhz19
# homeassistant.components.sensor.serial_pm
pmsensor==0.4

# homeassistant.components.prometheus
prometheus_client==0.2.0

# homeassistant.components.notify.pushbullet
# homeassistant.components.sensor.pushbullet
pushbullet.py==0.11.0

# homeassistant.components.canary
py-canary==0.5.0

# homeassistant.components.media_player.blackbird
pyblackbird==0.5

# homeassistant.components.deconz
pydeconz==42

# homeassistant.components.zwave
pydispatcher==2.0.5

# homeassistant.components.litejet
pylitejet==0.1

# homeassistant.components.media_player.monoprice
pymonoprice==0.3

# homeassistant.components.alarm_control_panel.nx584
# homeassistant.components.binary_sensor.nx584
pynx584==0.4

# homeassistant.auth_providers.modules.totp
# homeassistant.components.sensor.otp
pyotp==2.2.6

# homeassistant.components.qwikswitch
pyqwikswitch==0.8

# homeassistant.components.sensor.darksky
# homeassistant.components.weather.darksky
python-forecastio==1.4.0

# homeassistant.components.nest
python-nest==4.0.3

# homeassistant.components.sensor.whois
pythonwhois==2.4.3

# homeassistant.components.tradfri
pytradfri[async]==5.5.1

# homeassistant.components.device_tracker.unifi
pyunifi==2.13

# homeassistant.components.upnp
pyupnp-async==0.1.0.2

# homeassistant.components.notify.html5
pywebpush==1.6.0

# homeassistant.components.python_script
restrictedpython==4.0b4

# homeassistant.components.rflink
rflink==0.0.37

# homeassistant.components.ring
ring_doorbell==0.2.1

# homeassistant.components.media_player.yamaha
rxv==0.5.1

# homeassistant.components.sleepiq
sleepyq==0.6

# homeassistant.components.climate.honeywell
somecomfort==0.5.2

# homeassistant.components.recorder
# homeassistant.scripts.db_migrator
# homeassistant.components.sensor.sql
sqlalchemy==1.2.9

# homeassistant.components.statsd
statsd==3.2.1

# homeassistant.components.camera.uvc
uvcclient==0.10.1

# homeassistant.components.config.config_entries
voluptuous-serialize==1

# homeassistant.components.vultr
vultr==0.1.2

# homeassistant.components.wake_on_lan
# homeassistant.components.media_player.panasonic_viera
# homeassistant.components.media_player.samsungtv
# homeassistant.components.switch.wake_on_lan
wakeonlan==1.0.0

# homeassistant.components.cloud
warrant==0.6.1

# homeassistant.components.sensor.yahoo_finance
yahoo-finance==1.4.0
